[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CPUSETS_V1 [=n]
 Type  : bool
 Defined at init/Kconfig:1146
   Prompt: Legacy cgroup v1 cpusets controller
   Depends on: CGROUPS [=y] && CPUSETS [=y]
   Location:
     -> General setup
       -> Control Group support (CGROUPS [=y])
         -> Cpuset controller (CPUSETS [=y])
           -> Legacy cgroup v1 cpusets controller (CPUSETS_V1 [=n])

 Commit: 1abab1ba0775 (cgroup/cpuset: guard cpuset-v1 code under CONFIG_CPUSETS_V1)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
CONFIG_CPUSETS_V1 enable as a builtin.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>