[redhat] New configs in mm/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SLUB_RCU_DEBUG [=y]
 Type  : bool
 Defined at mm/Kconfig.debug:73
   Prompt: Enable UAF detection in TYPESAFE_BY_RCU caches (for KASAN)
   Depends on: SLUB_DEBUG [=y] && KASAN [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> Enable SLUB debugging support (SLUB_DEBUG [=y])
           -> Enable UAF detection in TYPESAFE_BY_RCU caches (for KASAN) (SLUB_RCU_DEBUG [=y])

 Commit: b8c8ba73c68b (slub: Introduce CONFIG_SLUB_RCU_DEBUG)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_SLUB_RCU_DEBUG disabled for rhel/generic, enabled for
  rhel/debug.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>