[redhat] gitlab-ci.yml: trigger bot pipelines on source project & branch name
Currently we trigger the limited srpm-only pipelines based on the
pipeline author username. This works fine when the MR is first opened,
but if anybody needs to trigger a new pipeline on these MRs, then they
are the pipeline author and we get a full set of pipelines instead of
the limited set.

Improve this by instead matching on the MR's source branch project ID
and branch name.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>