[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_QDMA [=n]
 Type  : tristate
 Defined at drivers/dma/amd/Kconfig:3
   Prompt: AMD Queue-based DMA
   Depends on: DMADEVICES [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
         -> AMD Queue-based DMA (AMD_QDMA [=n])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=y] && REGMAP_MMIO [=n]

 Commit: 73d5fc92a11c (dmaengine: amd: qdma: Add AMD QDMA driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>