[redhat] New configs in drivers/media
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: USB_EXTRON_DA_HD_4K_PLUS_CEC [=n]
 Type  : tristate
 Defined at drivers/media/cec/usb/extron-da-hd-4k-plus/Kconfig:2
   Prompt: Extron DA HD 4K Plus CEC driver
   Depends on: MEDIA_CEC_SUPPORT [=y] && USB_SUPPORT [=y] && TTY [=y] && VIDEO_DEV [=m] && USB [=y] && USB_ACM [=m]
   Location:
     -> Device Drivers
       -> CEC support
         -> HDMI CEC drivers (MEDIA_CEC_SUPPORT [=y])
           -> Extron DA HD 4K Plus CEC driver (USB_EXTRON_DA_HD_4K_PLUS_CEC [=n])
 Selects: CEC_CORE [=m] && SERIO [=y] && SERIO_SERPORT [=m]

 Commit: 056f2821b631 (media: cec: extron-da-hd-4k-plus: add the Extron DA HD 4K Plus CEC driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>