[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PCI_NPEM [=n]
 Type  : bool
 Defined at drivers/pci/Kconfig:146
   Prompt: Native PCIe Enclosure Management
   Depends on: PCI [=y] && LEDS_CLASS [=y]=y [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> Native PCIe Enclosure Management (PCI_NPEM [=n])

 Commit: 4e893545ef87 (PCI/NPEM: Add Native PCIe Enclosure Management support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>