[redhat] New configs in drivers/input
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MOUSE_PS2_PIXART [=y]
 Type  : bool
 Defined at drivers/input/mouse/Kconfig:72
   Prompt: PixArt PS/2 touchpad protocol extension
   Depends on: INPUT [=y] && INPUT_MOUSE [=y] && MOUSE_PS2 [=y]
   Visible if: INPUT [=y] && INPUT_MOUSE [=y] && MOUSE_PS2 [=y] && EXPERT [=y]
   Location:
     -> Device Drivers
       -> Input device support
         -> Generic input layer (needed for keyboard, mouse, ...) (INPUT [=y])
           -> Mice (INPUT_MOUSE [=y])
             -> PS/2 mouse (MOUSE_PS2 [=y])
               -> PixArt PS/2 touchpad protocol extension (MOUSE_PS2_PIXART [=y])

 Commit: 740ff03d7238 (Input: Add driver for PixArt PS/2 touchpad)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>