[redhat] New configs in drivers/extcon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: EXTCON_LC824206XA [=n]
 Type  : tristate
 Defined at drivers/extcon/Kconfig:78
   Prompt: LC824206XA extcon Support
   Depends on: EXTCON [=y] && I2C [=y] && POWER_SUPPLY [=y]
   Location:
     -> Device Drivers
       -> External Connector Class (extcon) support (EXTCON [=y])
         -> LC824206XA extcon Support (EXTCON_LC824206XA [=n])

 Commit: 9e1897cb9568 (extcon: Add LC824206XA microUSB switch driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>