[redhat] New configs in drivers/nvmem
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVMEM_LAYOUT_U_BOOT_ENV [=n]
 Type  : tristate
 Defined at drivers/nvmem/layouts/Kconfig:29
   Prompt: U-Boot environment variables layout
   Depends on: NVMEM [=y] && NVMEM_LAYOUTS [=y]
   Location:
     -> Device Drivers
       -> NVMEM Support (NVMEM [=y])
         -> Layout Types
           -> U-Boot environment variables layout (NVMEM_LAYOUT_U_BOOT_ENV [=n])
 Selects: CRC32 [=y] && GENERIC_NET_UTILS [=y]
 Selected by [n]:
   - NVMEM_U_BOOT_ENV [=n] && NVMEM [=y] && OF [=y] && MTD [=m]

 Commit: 5f15811286af (nvmem: layouts: add U-Boot env layout)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>