Trim Changelog for 6.12
Every time we rebase, the changelog gets completely confused, and the
first merge creates an entire duplication of the changelog. This has to
be manually trimmed so we don't end up with multiple copies of
everything. I failed to do this during the merge window, so I am doing
it now. End result is the same.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>