acpi/x86: s2idle: rename Screen On/Off to Display On/Off
Microsoft and Intel use the term "Display" to refer to the _DSM 3,4
calls and the term "Screen" for the state ("Screen Off"). Currently,
the code uses "Screen On/Off" to name the variables, which is about
to become confusing as they become callbacks. To prepare for that,
rename the variables to "Display On/Off".

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>