acpi/x86: s2idle: rename MS Exit/Entry to Sleep Exit/Entry
Microsoft refers to the _DSMs 7,8 as "Sleep Entry" and "Sleep Exit".
Currently, the code uses "MS Entry/Exit" to name the variables, which is
confusing as it could either mean "Modern Standby" or "Microsoft" and is
not representative of the state. Rename as part of converting it into
a transition.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>