[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: X86_BUS_LOCK_DETECT [=y]
 Type  : bool
 Defined at arch/x86/Kconfig:2431
   Prompt: Split Lock Detect and Bus Lock Detect support
   Depends on: CPU_SUP_INTEL [=y] || CPU_SUP_AMD [=y]
   Location:
     -> Processor type and features
       -> Split Lock Detect and Bus Lock Detect support (X86_BUS_LOCK_DETECT [=y])

 Commit: 350afa8a1101 (x86/split_lock: Move Split and Bus lock code to a dedicated file)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>