[redhat] New configs in drivers/gpio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: GPIO_MPSSE [=n]
 Type  : tristate
 Defined at drivers/gpio/Kconfig:1857
   Prompt: FTDI MPSSE GPIO support
   Depends on: GPIOLIB [=y] && USB [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> USB GPIO expanders
           -> FTDI MPSSE GPIO support (GPIO_MPSSE [=n])
 Selects: GPIOLIB_IRQCHIP [=y]

 Commit: c46a74ff05c0 (gpio: add support for FTDI's MPSSE as GPIO)

---

 Symbol: GPIO_POLARFIRE_SOC [=n]
 Type  : bool
 Defined at drivers/gpio/Kconfig:547
   Prompt: Microchip FPGA GPIO support
   Depends on: GPIOLIB [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> GPIO Support (GPIOLIB [=y])
         -> Memory mapped GPIO drivers
           -> Microchip FPGA GPIO support (GPIO_POLARFIRE_SOC [=n])
 Selects: REGMAP_MMIO [=n]

 Commit: a987b78f3615 (gpio: mpfs: add polarfire soc gpio support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>