[redhat] New configs in drivers/hid
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HID_KYSONA [=n]
 Type  : tristate
 Defined at drivers/hid/Kconfig:471
   Prompt: Kysona devices
   Depends on: HID_SUPPORT [=y] && HID [=y] && USB_HID [=y]
   Location:
     -> Device Drivers
       -> HID bus support (HID_SUPPORT [=y])
         -> HID bus core support (HID [=y])
           -> Special HID drivers
             -> Kysona devices (HID_KYSONA [=n])

 Commit: 30c32d052945 (HID: Kysona: Add basic battery reporting for Kysona M600)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>