[redhat] New configs in drivers/misc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MCHP_LAN966X_PCI [=n]
 Type  : tristate
 Defined at drivers/misc/Kconfig:613
   Prompt: Microchip LAN966x PCIe Support
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> Misc devices
         -> Microchip LAN966x PCIe Support (MCHP_LAN966X_PCI [=n])
 Selects: OF [=n] && OF_OVERLAY [=n] && IRQ_DOMAIN [=y]

 Commit: 185686beb464 (misc: Add support for LAN966x PCI device)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>