[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: FUNCTION_GRAPH_RETADDR [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:245
   Prompt: Kernel Function Graph Return Address
   Depends on: FTRACE [=y] && FUNCTION_GRAPH_TRACER [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Kernel Function Tracer (FUNCTION_TRACER [=y])
           -> Kernel Function Graph Tracer (FUNCTION_GRAPH_TRACER [=y])
             -> Kernel Function Graph Return Address (FUNCTION_GRAPH_RETADDR [=n])

 Commit: 21e92806d39c (function_graph: Support recording and printing the function return address)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>