[redhat] New configs in net/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_NET_SMALL_RTNL [=n]
 Type  : bool
 Defined at net/Kconfig.debug:28
   Prompt: Add extra per-netns mutex inside RTNL
   Depends on: DEBUG_KERNEL [=y] && NET [=y] && LOCK_DEBUGGING_SUPPORT [=y]
   Location:
     -> Kernel hacking
       -> Networking Debugging
         -> Add extra per-netns mutex inside RTNL (DEBUG_NET_SMALL_RTNL [=n])
 Selects: PROVE_LOCKING [=n]

 Commit: 76aed95319da (rtnetlink: Add per-netns RTNL.)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_DEBUG_NET_SMALL_RTNL enable for rhel debug.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>