[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HOTPLUG_PCI_OCTEONEP [=n]
 Type  : bool
 Defined at drivers/pci/hotplug/Kconfig:121
   Prompt: Marvell OCTEON PCI Hotplug driver
   Depends on: PCI [=y] && HOTPLUG_PCI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> Support for PCI Hotplug (HOTPLUG_PCI [=y])
           -> Marvell OCTEON PCI Hotplug driver (HOTPLUG_PCI_OCTEONEP [=n])

 Commit: e434e54d3ffc (PCI: hotplug: Add OCTEON PCI hotplug controller driver)

---

 Symbol: PCIE_TPH [=n]
 Type  : bool
 Defined at drivers/pci/Kconfig:176
   Prompt: TLP Processing Hints
   Depends on: PCI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> TLP Processing Hints (PCIE_TPH [=n])

 Commit: f69767a1ada3 (PCI: Add TLP Processing Hints (TPH) support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>