[redhat] New configs in lib/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TEST_KALLSYMS [=n]
 Type  : tristate
 Defined at lib/Kconfig.debug:2929
   Prompt: module kallsyms find_symbol() test
   Depends on: RUNTIME_TESTING_MENU [=y] && m [=m] && MODULES [=y]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> Runtime Testing (RUNTIME_TESTING_MENU [=y])
           -> module kallsyms find_symbol() test (TEST_KALLSYMS [=n])
 Selects: TEST_RUNTIME [=n] && TEST_RUNTIME_MODULE [=n] && TEST_KALLSYMS_A [=n] && TEST_KALLSYMS_B [=n] && TEST_KALLSYMS_C [=n] && TEST_KALLSYMS_D [=n]

 Commit: 84b4a51fce4c (selftests: add new kallsyms selftests)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>