[redhat] New configs in drivers/usb
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: TYPEC_MUX_TUSB1046 [=n]
 Type  : tristate
 Defined at drivers/usb/typec/mux/Kconfig:69
   Prompt: TI TUSB1046 Type-C crosspoint switch driver
   Depends on: USB_SUPPORT [=y] && TYPEC [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> USB support (USB_SUPPORT [=y])
         -> USB Type-C Support (TYPEC [=y])
           -> USB Type-C Multiplexer/DeMultiplexer Switch support
             -> TI TUSB1046 Type-C crosspoint switch driver (TYPEC_MUX_TUSB1046 [=n])

 Commit: 7c561b8c3e83 (usb: typec: mux: Add support for the TUSB1046 crosspoint switch)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>