[redhat] New configs in arch/arm64
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: HISILICON_ERRATUM_162100801 [=y]
 Type  : bool
 Defined at arch/arm64/Kconfig:1238
   Prompt: Hip09 162100801 erratum support
   Location:
     -> Kernel Features
       -> ARM errata workarounds via the alternatives framework
         -> Hip09 162100801 erratum support (HISILICON_ERRATUM_162100801 [=y])

 Commit: f82e62d470cc (irqchip/gicv3-its: Add workaround for hip09 ITS erratum 162100801)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>