configs: Enable CONFIG_NETKIT for RHEL
The Netkit virtual device is a light-weight alternative to veth with
native TC-BPF program support. Enabling this is in line with our goal of
making RHEL a first-class platform for using BPF on top of.

Since the config is now the same for Fedora and RHEL, move it into
common/generic.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>