aarch64: Options required for UEFI HTTP boot on SystemReady-DT
To support UEFI HTTP(S) boot on SystemReady-DT compliant
devices we need PMEM support to be built into the kernel
so that the memory/storage options are available in kernel
early on to be able to start the install. The SystemReady-DT
platforms are supported on both RHEL and Fedora for a
number of usecases.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>