Input: xpad - add support for TECNO Pocket Go
TECNO Pocket Go is a kickstarter handheld by
manufacturer TECNO Mobile. It poses a unique
feature: it does not have a display. Instead,
the handheld is essentially a pc in a
controller. As customary, it has an xpad
endpoint, a keyboard endpoint, and a vendor
endpoint for its vendor software.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>