ACPI: s2idle: Trigger _WAK as part of exiting s2idle
Currently, _WAK is not called when exiting s2idle. It is only called
when exiting states S3 and S4. The ACPI spec has not been updated for
S0ix, which makes things more ambiguous.

Certain OneXPlayer and GPD devices contain a hibernation feature where
the EC wakes up the device and pretends to overheat, causing Windows to
hibernate. This is a three part process, where after the device enters
S2idle through calling any _DSM callback (e.g., Sleep Entry, LPS0
Entry), they begin a counter that measures how much battery is drained.
Once it is at 5%, the EC of the device wakes it up if its in LPS0 by
pressing the powerbutton, and once it is in Modern Standby again, it
triggers the overheat event.

The problem here is that the notification that the system woke up was
placed in the outer part of _WAK, applying to exiting any SX state.
However, without it being called from s2idle, the EC never calls the
underlying event, leaving us with a powerbutton wake-up that is not
user initiated. By calling _WAK in the same place it is called after
S3, this problem is solved which in parallel shows that Windows does
indeed call _WAK during the exit from LPS0.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>