redhat/Makefile: Fix long dist-full-help execution time
@jstancek and @dzickus both noted that the execution time of
dist-full-help was longer than eight seconds.  This happens due to the
long lookup times of the git merge-base command in the localversion
variables code.

The help commands do not need 99.9% of the variable definitions, so wrap
the variable definitions with an ifeq statement to avoid execution for the
help commands.  The include of Makefile.cross and definition of HELP_TYPES
are required by the help commands, so those do remain included.

The dist-git and dist-git-headers must be modified to make the checks for
RHDISTGIT_BRANCH variable local instead of global.

Fix the long dist-full-help execution time by skipping most of the
variable definition code.

Reported-by: Jan Stancek <jstancek@redhat.com>
Reported-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Prarit Bhargava <prarit@redhat.com>