redhat/configs: automotive: set CONFIG_TIMEOUT_PANIC
For automotive any condition leading to PANIC should be considered as
dangerous and fatal, hence we want to stop operation as soon as possible.
So set the panic timeout value < 0 to reboot immediately for the generic
kernel config.

For debug we might be interested on capture the panic an don't reboot
immediately, so maintain the current behaviour and wait forever to let
the user decide what to do.

Signed-off-by: Enric Balletbo i Serra <eballetb@redhat.com>