[redhat] New configs in lib/kunit
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KUNIT_AUTORUN_ENABLED [=y]
 Type  : bool
 Defined at lib/kunit/Kconfig:84
   Prompt: Default value of kunit.autorun
   Depends on: KUNIT [=m]
   Location:
     -> Kernel hacking
       -> Kernel Testing and Coverage
         -> KUnit - Enable support for unit tests (KUNIT [=m])
           -> Default value of kunit.autorun (KUNIT_AUTORUN_ENABLED [=y])

 Commit: 31691914c392 (kunit: Introduce autorun option)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>