[redhat] New configs in drivers/infiniband
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INFINIBAND_HNS_HIP08 [=n]
 Type  : tristate
 Defined at drivers/infiniband/hw/hns/Kconfig:2
   Prompt: Hisilicon Hip08 Family RoCE support
   Depends on: INFINIBAND [=m] && (INFINIBAND_USER_ACCESS [=m] || !INFINIBAND_USER_ACCESS [=m]) && !UML && (ARM64 [=y] || COMPILE_TEST [=n] && 64BIT [=y]) && PCI [=y] && HNS3 [=m]
   Location:
     -> Device Drivers
       -> InfiniBand support (INFINIBAND [=m])
         -> Hisilicon Hip08 Family RoCE support (INFINIBAND_HNS_HIP08 [=n])

 Commit: dd74282df573 (RDMA/hns: Initialize the PCI device for hip08 RoCE)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>