[redhat] New configs in drivers/iommu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ARM_SMMU_MMU_500_CPRE_ERRATA [=y]
 Type  : bool
 Defined at drivers/iommu/Kconfig:370
   Prompt: Enable errata workaround for CPRE in SMMU reset path
   Depends on: IOMMU_SUPPORT [=y] && ARM_SMMU [=y]
   Location:
     -> Device Drivers
       -> IOMMU Hardware Support (IOMMU_SUPPORT [=y])
         -> ARM Ltd. System MMU (SMMU) Support (ARM_SMMU [=y])
           -> Enable errata workaround for CPRE in SMMU reset path (ARM_SMMU_MMU_500_CPRE_ERRATA [=y])

 Commit: ef4144b1b47d (iommu/arm-smmu: Re-enable context caching in smmu reset operation)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>