[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: BD79703 [=n]
 Type  : tristate
 Defined at drivers/iio/dac/Kconfig:351
   Prompt: ROHM Semiconductor BD79703 DAC driver
   Depends on: IIO [=m] && SPI [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Digital to analog converters
           -> ROHM Semiconductor BD79703 DAC driver (BD79703 [=n])
 Selects: REGMAP_SPI [=m]

 Commit: af6aca656a85 (iio: dac: Support ROHM BD79703 DAC)

---

 Symbol: OPT4060 [=n]
 Type  : tristate
 Defined at drivers/iio/light/Kconfig:478
   Prompt: Texas Instruments OPT4060 RGBW Color Sensor
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Light sensors
           -> Texas Instruments OPT4060 RGBW Color Sensor (OPT4060 [=n])
 Selects: REGMAP_I2C [=m] && IIO_BUFFER [=y] && IIO_TRIGGERED_BUFFER [=m]

 Commit: 0c6db4506ad0 (iio: light: Add support for TI OPT4060 color sensor)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>