[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_AE4DMA [=n]
 Type  : tristate
 Defined at drivers/dma/amd/Kconfig:4
   Prompt: AMD AE4DMA Engine
   Depends on: DMADEVICES [=y] && (X86_64 [=y] || COMPILE_TEST [=n]) && PCI [=y] && AMD_PTDMA [=m]
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
         -> AMD AE4DMA Engine (AMD_AE4DMA [=n])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=y]

 Commit: 90a30e268d9b (dmaengine: ae4dma: Add AMD ae4dma controller driver)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_AMD_AE4DMA: enable as a module for x86.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>