[redhat] New configs in lib/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRC_OPTIMIZATIONS [=y]
 Type  : bool
 Defined at lib/Kconfig:243
   Prompt: Enable optimized CRC implementations
   Visible if: EXPERT [=y]
   Location:
     -> Library routines
       -> Enable optimized CRC implementations (CRC_OPTIMIZATIONS [=y])

 Commit: b0430f39de08 (lib/crc: simplify the kconfig options for CRC implementations)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>