acpi/x86: s2idle: Add AC/DC burst wakeup support
According to Microsoft, when a power capable dock/charger is
attached/detached to the device, the SOC is expected to wake up.
Further, when the event is a connection to AC, the display of the
system is expected to turn on for five seconds.

The absence of this feature causes an issue where the original Ally
unit will get stuck in prochot.

Add the ability for s2idle to wake up the system from sleep on adapter
changes and add it to the delay quirk table.

Co-developed-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>