fixup! [FOR-UPSTREAM] Bluetooth: btrtl: add suspend handling to ignore BT_DIS
Turns out 5/7 call chains that can lead to the close flushing the
request work come from another work item, most of which don't set any
state that we can gate the enabling on.

Given the number of possible paths that this can end up with a guranteed
lock inversion issue, it seems more prudent to rely on the same
guarantees that made the original btrtl code not need to lock, which
was to only do requests at start of day before registering with the
HCI subsystem so that no lock taking is ever required.

This commit simplifies the design considerably, and avoids any chance
of the lock inversion happing.

Signed-off-by: Robert Beckett <bob.beckett@collabora.com>