ACPI: platform_profile: add all platform profiles to amd-pmf
amd-pmf is currently the only (sometimes) secondary consumer of the
platform-profiles ABI. This is what prompted this elaborate refactor
after all. However, it seems that when it is active with a WMI driver
that does not implement the same name of the power saving variant
(e.g., COOL instead of LOW_POWER), it eats that option.

This causes increased power consumption when userspace software
such as ppd or tuned tries to write to the unified platform-profile,
as it will no longer find a low power profile.

As amd-pmf for most laptops is a secondary consumer, make sure it
implements all the platform profiles, so regardless of what consumers
write to, it is available.

Tested on an Asus Z13.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>