platform/x86: oxpec: Add support for OneXPlayer G1
The OneXPlayer G1 is a new clamshell formfactor by OneXPlayer.
It has the same registers as an OneXPlayer X1.

However, on this one the turbo register is multiplexed and contains an
extra bit (0x02) which controls the fan function. Therefore, we need to
only change the 0x40 bit to not affect other behavior. Collapse the
disable and enable functions, and apply the fix to all devices.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>