platform/x86: oxpec: Make turbo val apply a bitmask
On the OneXPlayer G1, the turbo register is multiplexed and contains an
extra bit (0x02) which controls the fan function (always 0x02 on boot).
Therefore, we need to only change the 0x40 bit to not affect other
behavior. Collapse the disable and enable functions, and apply the fix
to all devices.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>