[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_HTU31 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:802
   Prompt: Measurement Specialties HTU31 humidity and temperature sensor
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> Measurement Specialties HTU31 humidity and temperature sensor (SENSORS_HTU31 [=n])
 Selects: CRC8 [=m]

 Commit: bf1bb26f23f1 (hwmon: add driver for HTU31)

---

 Symbol: SENSORS_INA233 [=n]
 Type  : tristate
 Defined at drivers/hwmon/pmbus/Kconfig:136
   Prompt: Texas Instruments INA233 and compatibles
   Depends on: HWMON [=y] && PMBUS [=m]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> PMBus support (PMBUS [=m])
           -> Texas Instruments INA233 and compatibles (SENSORS_INA233 [=n])

 Commit: b64b6cb163f1 (hwmon: Add driver for TI INA233 Current and Power Monitor)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>