[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_MAX77705 [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:919
   Prompt: Maxim MAX77705 PMIC Support
   Depends on: HAS_IOMEM [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Maxim MAX77705 PMIC Support (MFD_MAX77705 [=n])
 Selects: MFD_CORE [=y] && MFD_SIMPLE_MFD_I2C [=n]

 Commit: c8d50f029748 (mfd: Add new driver for MAX77705 PMIC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>