[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PINCTRL_TLMM_TEST [=n]
 Type  : tristate
 Defined at drivers/pinctrl/qcom/Kconfig.msm:440
   Prompt: Qualcomm TLMM test driver
   Depends on: PINCTRL [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && PINCTRL_MSM [=y] && (ARM64 [=y] || COMPILE_TEST [=n]) && KUNIT [=m]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
         -> Qualcomm core pin controller driver (PINCTRL_MSM [=y])
           -> Qualcomm TLMM test driver (PINCTRL_TLMM_TEST [=n])

 Commit: c7984dc0a2b9 (pinctrl: qcom: Add test case for TLMM interrupt handling)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>