drm/amd: Use suspend and hibernate post freeze notifications
commit 2965e6355dcd ("drm/amd: Add Suspend/Hibernate notification
callback support") introduced a VRAM eviction earlier in the PM
sequences when swap was still available for evicting to. This helped
to fix a number of memory pressure related bugs but also exposed a
new one. If a userspace process is actively using the GPU when suspend
starts then a deadlock could occur. For this reason, it was reverted.

Re-introduce it and instead of going off the prepare notifier, use the
PM notifiers that occur after processes have been frozen to do evictions.

Closes: https://gitlab.freedesktop.org/drm/amd/-/issues/4178
Fixes: 2965e6355dcd ("drm/amd: Add Suspend/Hibernate notification callback support")
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>