redhat/configs: Add configs for new ov02c10 and ov02e10 drivers
Add configs for the just added ov02c10 and ov02e10 drivers. This follows
how the config for e.g. the already existing ov01a10 driver is handled.

On Fedora the new drivers are build as module for all architectures,
because e.g. various aarch64 laptops also use one of these 2 sensors
for their cameras.

On RHEL the drivers are disabled by default and only enabled as module in
x86 builds, so that the cameras in recent Dell laptop models will work.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>