platform/x86: msi-wmi-platform: Add tunable support via firmware attributes
Adds PL1/PL2/SPL/SPPT/FPPT support through the firmware attributes
interface. The min and max values are quirked, and the attributes are
only defined if they are set to a non-zero value. These values are meant
to be set in conjunction with shift mode, where shift mode automatically
sets an upper bound those (e.g., low-power would be used with 8W).

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>