[redhat] New configs in samples/damon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SAMPLE_DAMON_MTIER [=n]
 Type  : bool
 Defined at samples/damon/Kconfig:30
   Prompt: DAMON sample module for memory tiering
   Depends on: SAMPLES [=y] && DAMON [=y] && DAMON_PADDR [=y]
   Location:
     -> Kernel hacking
       -> Sample kernel code (SAMPLES [=y])
         -> DAMON Samples
           -> DAMON sample module for memory tiering (SAMPLE_DAMON_MTIER [=n])

 Commit: 82a08bde3cf7 (samples/damon: implement a DAMON module for memory tiering)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>