[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_OC_WATCHDOG [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:1375
   Prompt: Intel OC Watchdog
   Depends on: WATCHDOG [=y] && (X86 [=y] || COMPILE_TEST [=n]) && ACPI [=y] && HAS_IOPORT [=y]
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
         -> Intel OC Watchdog (INTEL_OC_WATCHDOG [=n])
 Selects: WATCHDOG_CORE [=y]

 Commit: 535d1784d8a9 (watchdog: Add driver for Intel OC WDT)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>