redhat/spec: Disable gdb index for riscv cross-compile
The current version of gdb-minimal in CentOS Stream 10 crashes when
find-debuginfo tries to build the gdb index for a riscv64 kernel. Until
we update GDB, this temporary measure should help prevent CI pipelines
for c10s from failing on the riscv64 build.

This issue does not affect Fedora, ELN, or native builds on a riscv64
host running the RHEL 10 RISC-V Developer Preview, and is expected to be
fixed with GDB 16.

Signed-off-by: Jennifer Berringer <jberring@redhat.com>