platform/x86: asus-wmi: Fix WLAN button mapping on ASUS ROG Z13
On 9286dfd5735b ("platform/x86: asus-wmi: Fix spurious rfkill on
UX8406MA"), Mathieu adds a quirk for the Zenbook Duo to ignore
the code 0x5f (WLAN button disable). On that laptop, this code
is triggered when the device keyboard is attached.

On the ASUS ROG Z13 2025, this code is triggered when pressing the side
button of the device, which is used to open Armoury Crate in Windows.

As this is becoming a pattern, where newer Asus laptops use this keycode
for emitting events, let's convert the wlan ignore quirk to instead
allow emitting codes, so that userspace programs can listen to it and
so that it does not interfere with the rfkill state.

With this patch, the Z13 wil emit KEY_PROG3 and the Duo will remain
unchanged and emit no event. As a reference, the Z13 emits 0x75 when its
keyboard is attached. Therefore, if the Duo is to emit a keycode, it
should be introduced in a separate patch alongside the Z13.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>